requirements.txt
# Usage: pip3 install -r requirements.txt

bcrypt ~= 3.1.7
gevent ~= 21.8.0
gunicorn ~= 20.1.0
Flask ~= 2.0.1
Flask-SocketIO ~= 5.1.1
mysql-connector-python ~= 8.0.19
pytest ~= 6.2.1
